Added new util: "renderTimeSpan"
This takes in a startTimestamp and endTimestamp, and performs the logic
to determine how to display it:

- Just a start date & time (point-event with time)
- Full day span (only display dates)
- Custom timespan (includes time)

We need this anywhere that requires displaying a timespan, and it’s
good to have this be consistent across the app. I won’t implement this
across the app until we’ve confirmed this works as expected, with every
combination.

#248