Temp fix for re-drawing chart when new metric values received
This was something that worked fine at some point, but as we increased
the complexity of the chart component, we forgot to redraw the chart
when the `@props.metricValues` changed.

This is a temporary (slow) fix till we figure out a better C3 chart
implementation (#833) which will reactively handle all the updates for
us - without having to completely re-draw everything every time.

FYI @conceptualspace @jophis