use 3d transform to hide/show prognotes pane
> with a large client file, there was significant latency (up to a few
seconds) every time the user would navigate to or from the prognotes
pane via the sidebar
> profiling revealed that the delay was entirely the result of
recalculating style via showwhen (changing the display attribute)
> it turns out that with ~ 300 prognotes and other data, the prognotes
component ends up a with massive 20,000 nodes in the DOM to be modified,
hence the latency
> so instead of hiding and showing the prognotes tab, forcing the
reflow, what if we could hide it in such a way that the dom stayed in
tact? we can scale to 0 with a css transform. it still leaves a big
blank space on the page when it is "hidden" this way, so to compensate
for this, negative margins are calculated on the other pages -- their
content is just displayed overtop of the hidden prognotes page
> the result is smooth navigation in < 200ms
> its a hack, but it works? comments welcome
@applebya @jophis