removed tooltips on template/print buttons.
They were causing more trouble than they’re worth and seemed out of
place most of the time. Maybe adding some text to the button is the way
to go. Either way, the template button will not be used by many, and
its immediately clear whats it’s purpose is upon clicking it. And the
print icon is self explanatory.