Improve hover/highlighting perf on analysisTab eventTypes
I was pretty disappointed with the performance of hovering over an
eventType to highlight it on the graph - especially with the test data
that David plans on using.

It really doesn’t require re-rendering the entire analysisTab component
- only the InlineHighlightStyles itself.

So all I did was turn InlineHighlightStyles into a proper component,
with the highlightedEventTypeId handled internally as state. Methods
`add` and `remove` are available upstream through its ref.

#817