Custom shouldComponentUpdate for ProgNoteContainer
Since the attachments factor only solved the perf issue part-way, I
wrote this custom `shouldComponentUpdate` function.

Basically, we know that progEvents and attachments might be modified
(so `Imm.is`) whereas `progNoteHistory` can only grow inside, so no
need for the more expensive Imm.is there.

With a clientFile that has 1000 progNotes, it previously took about 5s
- whereas now it’s about 500ms .

FYI @conceptualspace

Related to conversation in #787