Remove maxDate constraint for full day
When creating a new progEvent, I wasn’t able to select a date later
than the current day. This is because a full-day timespan technically
has an endDate, so it was setting the datetimepicker’s maxDate to that
simulated endDate.

I copied over a bit of logic to check to see whether this is a full-day
timespan, and if so, don’t apply the maxDate.

Since the user can switch from a multi-dayspan back to a full-day, the
maxDate also needs to be removed in this case.

As stated in the TODO, this will eventually be refactored to something
more declarative when we add self-contained datetimepicker react
components. At least this plugs the hole :)