Chart now functional when only events selected
#889 #726

My idea was to fool the chart into thinking it always has a metric
selected, even when none are actually selected. I do this by injecting
a “hidden” metric into the top of the `_generateChart()` method, and
adding exceptions for it when needed.

The value of this hidden metric is -9999. The chart itself only goes
from 0-1 (or 0-2 when events are visible), so we’re assured it’s
sitting wayyyy outside of the chart area. This also means I don’t even
have to hide it with CSS, but it’s got a custom class just in case.

When we go to `_refreshSelectedMetrics()`, the hidden metric is added
back immediately after the chart’s cleared.