responsive layout for createClientDialog
#893 @applebya @conceptualspace finally happy with where this is at and
learned some more style along the way.
on a fresh install, the dialog is a single panel (no program or
template dropdown)
if one of programs or templates exist, they are place at the bottom of
the single panel.
if both programs and templates exist, they populate a second panel.