Move scrolling behaviour to `History` component, fix inf.loop
#896

Ok, so first I thought this was a problem with the stuff in the top of
the render getting out of sync, because maybe `@props` changed halfway
through. So I saved these to instance vars. Didn’t turn out being the
solution, but it’s a good refactor anyway.

The actual cause: componentDidUpdate attached the scroll listener every
time state/props changed, which would eventually add up - causing the
infinite loop.

What I did was move the scrolling logic to a History component, which
could reliably attach the scroll listener on componentDidMount (once).

historyCount methods are passed in from detailView. It’s internal
resetScroll is triggered via refs.

@conceptualspace