Use class vars for secondPass data, setState once finished
#903 @conceptualspace @jophis

- Axed the shouldComponentUpdate
- Class vars are set up on componentWillMount
- `@progNoteTotal` set in initial data-load
- Moved `_secondPass()` to below `_renewAllData()`, just makes it
easier to see the order of execution
- `_secondPass()` changes, as detailed in #903 discussion. Now saves
all secondPass stuff to class vars :)

We might be able to refactor this to a simpler`Async.until`, and keep
all these vars internal to _secondPass , but I doubt it’d make much of
a different perf-wise. Save for second pass at secondPass :P

I do hope we can fix the impure render issues, so we can let these
batches of progNotes trickle down to progNotesTab without causing such
a render-ruckus. In the meantime, it would be nice to have some
indication of progress - even just a spinner.