Fix display of selectedProgram in ManageAccountDialog
The program selection for ManageAccountDialog was a bit broken - just
wouldn’t display the currently selected program, but the actual
functionality still worked.

This resulted from an earlier fix where a distorted version of the
clientFile was passed in here (included program property), so it puked
on saving to DB. But now with the original userAccount data, we have to
figure out the program membership again. I need to be more careful with
testing every feature after fixing a component’s props.

@conceptualspace We might want to patch this into Griffin, or do
another mid-release soon with this and other minor fixes. Ugh, I know :/