Add markjs dep to /lib
#135

Ok, so I tried a whole bunch of different methods to accomplish this
highlighting behaviour by search term.

- Replacing actual text in data with `<span>…</span>` (too much dev
overhead for future displays/components, delays actual render)
- Single function to do this with jQuery (tried 2 of these, wouldn’t
work dependably)

I eventually landed on markjs. It’s only 13.4KB, so a very lightweight
dependancy - which doesn’t require jQuery. Seems to have a very active
community behind it.

The cost of another dep in this case appears to be worth it, but I’m
open to other solutions. I just like how it doesn’t really **need** the
render in order to function, so we don’t have to worry about mutating
the data itself - just the DOM after the fact.