Bring filtering (no value) logic to top-level, pass down filteredProgNote
#135

The keyword search feature requires that it only search through stuff
that has a value. This was being handled at the bottom-level (per
progNote), which made this impossible.

So, I’ve moved this filter logic up to the top-level, and the
filteredProgNote is passed down. It’s eventually used by ProgNoteView.

The property ‘data’ is now ignored in the search as well, so it only
searches the filteredProgNote, and any prog/globalEvents.

Also fixes some of the ordering logic that I broke earlier.

Some minor refactors all around.

Next up: Figure out why the hell it won’t filter globalEvents… (which
probably don’t need to be included in the entry data anyway)