Non-editable progEventWidget uses plain DIV for display
#135

Without this, I wasn’t able to get Markjs to highlight a number value,
since all metrics were using a disabled `<input>` to display the value.

We know the value is going to exist, or use the ‘—‘ placeholder.

Turned out to be way easier (and less risky) to fix this than I thought
:)