Only run AnalysisTab shallowCompare when the tab is visible
@conceptualspace @jophis

This provides a pretty significant perf boost to the application.

AnalysisTab doesn’t need to re-render every time we do something that
affects its props, since it’s not visible all of the time. Hell, the
user may not even open the tab at all.

I did a quick perf check on the render, for modifying a progNote.
Previously took 293ms, and now 238ms - so about 15-20% better. I think
this would be even more significant for a large dataset, since the
AnalysisTab render is damn expensive.

I’m not sure whether it’s worth doing something similar for
progNotesTab and planTab, since it’s likely the user will be switching
between the two more frequently - so good to have it already prepared.

This also solves the issue I pointed out with the stretched chart (from
today’s e-mail), since it now has a visible container to determine its
width from.