Forward hasChanges call to progNotesTabUi
#135 @conceptualspace

This happened because of the new container component. ClientFilePage
was expecting it to still have a hasChanges method, which is now down
in ProgNotesTabUi.

Added a much clearer error message for when `suggestClose()` fails.

Part of me thinks that we shouldn’t try/catch errors here, since it
should always work. Then when it does fail, we get an actual crash.