Separate filterCount to limit # displayed entries (while isFiltering)
#135

`@state.filterCount` fulfills similar duties as `historyCount`, but
just for while we’re in filter-mode (`isFiltering`).

The number is reset back to 10 when FilterBar opens.

Yet another condition to call on (what is now)
`_redrawSearchHighlighting()`. This is starting to feel a wee bit
redundant, so probably some perf to gain in this area.