Pass down dataFilterType to selectively hide unsearched data
#512

Passed down dataFilterType to each entry view component, so they know
whether or not to show data unrelated to the search.

This is kind of an an assumption on my part, that say the user only
wants to see the ‘targets’ when searching targets - and not progEvents
within the same note. It saves a heck of a lot of vertical space this
way, but on the other hand, they may still want to see the full
progNote.

Filtering by ‘events’ only shows progEvents and globalEvents.

Filtering by ‘targets’ doesn’t show progEvents or globalEvents

Filtering by ‘progNotes’ doesn’t show globalEvents