added search bar to metric manager layer.
#685 @applebya i added the search bar, but its not behaving as it
should. I think interfering with your custom style in
responsiveTable.styl.
I hacked it into place very inelegantly.. but it breaks down as the
window becomes smaller. maybe you can take a look at how to get the
‘table-tool-bar’ div to behave as it was intended.
I can get it working if i remove some of the other styl on the table,
namely the flex stuff. but that breaks other things.