EntriesListView comp contains unlimited-scroll & highlighting
#512

When filtering entries, our unlimited-scroll listener couldn’t know how
many entries was the maximum - so it had no way of distinguishing
whether the list was exhausted or not, which killed perf when scrolled
down to the bottom of the list.

Just like @conceptualspace did for detailView etc, all the list stuff
is now contained in a new container component: EntriesListView. It
handles all this stuff internally, and we know to redraw highlighting
when any of its props change (while isFiltering and has searchQuery).

Rather than explicitly pass all the props all over again to
ProgNoteContainer, `.extend` is wayyy more succinct and less
error-prone.