Conditionally show colorKeyBubble for entries/history
#742

FYI @conceptualspace @jophis my wifi is going to be spotty for the next
3-4d, so feel free to adjust things to your liking :)

ColorKeyBubble does not display when an authorProgram does not exist.

Entries View
- Moved colorKey to the far right. This is a bit more consistent with
how we do it in the history/revisions views. Also, it feels more
natural here pairing the authorProgram with the author userName.
- Rounded out the bubble a bit (fixed @ 10px)

Revisions/History View
- Consistent size of colorKeyBubble