Only highlight new search entries, progNotesList -> entriesListView
#956

Reviewer @conceptualspace

Much faster highlighting when unlimited-scrolling through results.

But if the filterCount is really big, and you then update the
search/filter parameters, we have to highlight everything all over
again. I think resetting the filterCount back to 10 every time we
update one of these parameters will solve this, and will be good for UX
too (so long as we scroll to top). That’s next up…