Reset filterCount and scroll when search params change
#956 @conceptualspace

(note the git diff here is messy, I only really added 1055-1061)

When the user changes the search/filter parameters, it makes sense that
we go back to our original filterCount of 10, so we don’t have to
re-highlight say 100 entries. This is fixes my last main perf concern.

It felt awkward not being back at the top of the results, so I have it
scrolling back to the top. I think is actually nice for UX - because
originally when the entry results got changed, suddenly we’d often be
looking at completely different notes while scrolled halfway down… it
was really disorientating.