Analysis slider only compares by startOf day, consistent w/ xTicks
#413

The slider component’s values are fed by xTicks, which are each always
the start of a given day. Zooming/panning the chart will likely provide
a timestamp anywhere within the given day, so the slider needs to know
that it’s still the same day - and will find the appropriate xTick.