Prevent redundant TimeSpanDate _onChange calls to parent
#413

When TimeSpanDate was updated by props, it’d always call `_onChange`
regardless of whether anything had actually changed or not. This would
call redundant re-renders to the parent, which slowed things down.

What would happen was this:
- Pick a new date in TimeSpanDate calendar popup, calls `_onChange`
- AnalysisTab re-renders, sends new props to TimeSpanDate
- TimeSpanDate would call `_onChange` a second time
- AnalysisTab renders again

A better fix would be messing with `componentDidUpdate`, but that could
have implications for setting min/max date - so I didn’t want to touch
it. I’d rather wait till we have a proper datetimepicker React
component to handle this stuff internally without all the hoopla (#914)