Chart doesn't use metricColors internally, fix red. _attachKeyBindings
Just some more perf work on analysisTab/chart - enhancements I figured
out while working on #413

- metricColors is handed up to analysisTab once the chart renders its
first time. Using the `onrendered` or `oninit` options in c3.generate
proves to be unreliable as to whether the `@_chart` var is available
yet. Rather, adding this below c3.generate works more better. We also
avoid a re-render of chart now, since it doesn’t need metricColors
internally.

- `_attachKeyBindings` was getting called a silly number of times,
since it was getting called by `_refreshProgEvents` and then internally
by the chart whenever it had any kind of a re-render. So these
listeners were stacking up. It might still be a bit leaky - the same
listener being stacked on to the same element - but this has already
improved the snappiness of hovering over an event.