Use renderLineBreaks in revisionHistory diff spans
#935

@conceptualspace for some reason, the `{newlineIsToken:true}`
diffTrimmedLines setting seems to be deleting all of the line-breaks
`\n` within the strings. I can presume this is intentional on their
part, but it takes away our ability to recognize line-breaks and
replace them with `<br>` tags.

I wonder if there’s a way for it to recognize line-breaks for the
purposes of diffing, but not remove them…?

Anyway, I’ve commented that setting out for the moment, which allowed
me to call `renderLineBreaks` in the value for each diff. I wanted to
just wait and call renderLineBreaks at the end (on the UI render), but
we’re already peppering the string with diff spans, and
renderLineBreaks is only made to handle a single string.

Also added empty string defaults for oldString & newString, just in
case it ever receives null/undefined :+1: