Update chart's internal zoom more aggressively, better perf
#413

Turns out that the debounced calling of `@_chart.zoom(domain)`
(workaround to fix C3’s internal memory of zoom) was causing some of
the bad behaviour - not the componentDidUpdate as I initially expected.

I’ve played with this quite a lot on my MacBook, and fine-tuned it to
the point it’s pretty hard to get much strange zooming behaviour when
rapidly zooming in & out.