Show metric definition when scale value within 0.025 of hoveredMetric
#784 #985

Went ahead with this, just to see how hard it’d be. Turns out, not bad
at all.

The main bit of logic added to `isHoveredMetric` is:
`Math.abs(@hoveredMetric.value - currentMetric.value) < 0.025`

I came up with 0.025 as the degree of similarity (scale is 0-1) while
fine-tuning this, to be close enough regardless of whether the metrics
chart is really tall or not.