Chart resizes to halfHeight when no metrics/events
#967

We call `@ _calculateChartHeight()` whenever the component updates
(before the chart gets re-rendered with any new data), and when the
window’s resized.

I tried this a few different ways, and this is by far the smoothest
solution. It absolutely needs to be resized first, otherwise the chart
often renders incorrectly (and would be worse on a slower machine).