Allow for chart's legend to control selectedMetricIds
This was pretty obvious, so I’ve added it in. You were previously able
to turn on/off a metric from the chart’s legend, but it wouldn’t
actually update selectedMetricIds for real.

We’ve discussed removing the legend entirely, which I’m starting to
feel might hurt the UX. Rather, it’d be nice for the legend item to
have a hoverTooltip with the metric definition.