Minimum of 0 is enforced in chart scaling
#998 @dgotlib reported that Gillian often has difficulty understanding
the magnitude of change conveyed in the chart, because our baseline is
the minimum recorded value.

For example, take the values of `[0.25, 0.5, 1.0]`. The magnitude of
change from 0.5 -> 1.0 is really 50%, but the chart would show a 66.66%
.

We discussed the issue that this causes for something like heart BPM -
which will create a good deal of empty space below the metric points
(and the slopes of data will be shallower). We resolved that the
tradeoff for better visual proportions is going to be more useful - and
for >v2 we can simply add a UI toggle to switch off the enforced 0
minimum (will create issue for this).

Added exception for negative minimum value, which would take precedent
over 0. This is a pretty rare case, but still possible (metric value
can be recorded as any +/- integer/decimal).