Custom scrollToElement util method
#994

I tried out a few different scrolling libraries for this first, just to
try and make something lightweight work. They turned out being a bit
bloated for our uses, and some weren’t even capable of scrolling within
a nested container.

Turned out finding this handy tutorial:
https://pawelgrzybek.com/page-scroll-in-vanilla-javascript/ , which I
bastardized the good stuff from.

The only issue with this solution is that `requestAnimationFrame` can
cause an infinite loop if we try and scroll past the maximum boundary,
so it can never reach the `destination` scrollTop. This `maxHeight`
calculation isn’t foolproof (yet), so I added a safeguard to ensure the
loop ends after the duration is reached.

We probably don’t need all of these `easings` options, but I’m just
trying them out for now to see what works. Maybe it’s worth keeping
them for the future, when we might need different easing for the
animation.