EntryDateNavigator component, calls EntriesListView.scrollToEntry()
#994

EntryDateNavigator lives in the FilterBar, but does not provide any
kind of filtering abilities (maybe we should rename to SearchBar?). Or,
we might consider having this persistently visible - whether or not
FilterBar is mounted.

It uses `@props.historyEntries` to populate the list of selectable
dates on the datetimepicker. These get updated as historyEntries change
(new progNote, event)

EntryDateNavigator is really just a smart button, which pops up the
datetimepicker when clicked. The first time it opens, the user can
click through the year -> month -> day for the fewest amount of clicks.

Once a date is selected, EntryDateNavigator calls
`EntriesListView.scrollToEntry` with the entry. We check to see whether
this entry is currently mounted in the DOM (index+1 is within
filterCount), and if it isn’t, we extend filterCount and wait to
performScroll() when it’s finished rendering the new entries.

I did a bit of extra styling to make the date selection more
visually-appealing, namely to add contrast between enabled & disabled
dates.