Re-init datetimepicker when historyEntries change, add safeguards
#994

I tried having a changed `@props.historyEntries` update
`@datetimepicker.enabledDates`, but there seems to be an upstream
problem here. The min/max needs to be updated too.

For now, a simple re-init works fine, though it’s a little expensive to
do every time historyEntries changes. At least this is only while
FilterBar is open though.

With the extra safeguards (and re-initing the datetimepicker), it won’t
try and scroll to a non-existent entry in historyEntries.