Add quick little animation to highlight scrollDestination entries
#994

As I used the navigate button more, I noticed that my eye doesn’t know
where to look after it finishes scrolling. I want to know at a quick
glance which entries I’ve scrolled to.

Without going crazy on colors here, I added a little .75 animation that
stretches out the entryHeader border.  It’s not too distracting, and
uses our existing color set.

Also fixed the issue with the border not normally stretching the full
way, because of our added margin on the right for selectedTarget
padding to not reflow the layout.