@setTimeSpanRange will subtract from the selected endDate
#970 @conceptualspace  span range buttons (week, day, 3 months etc)
will subtract the desired span from the SELECTED endDate instead of the
last date in the range of data.
previously, the buttons would snap the span to the end of the range of
data, now they will remain in the range that is currently selected. So
effectively, the buttons in the toolbar will only change the ‘start’
value of the timespan displayed.