Pass historyEntries, planTargetsById, and attachments to RevisionHistory
#777

Need this data passed down for when we call on ProgNoteContents for the
creation revision.

We work out attachments by pulling them from the historyEntry (which
already has its list of attachments figured out). Not exactly elegant
(or future-proof, say if you delete an attachment), but this works fine
for now.