Datetimepicker remains open after selection, can toggle or blur away
#994

With fresh eyes, I find the calendar closing (in addition to the
scrolling, and the spinner animation) to be a bit distracting. I might
even consider removing the spinner animation as well, and going with
something more subtle.

Anyhow, keeping the datetimepicker allows the user to skip around the
progNotes, without forcing them to keep re-opening it.

The datetimepicker gets hidden again if you click the button, or blur
away from it. We could enable the (x) close button, but that might be
overkill.

Overall, this flows quite a bit better now.