Fix metricWidget's inputWidth for many digits
FYI @jophis

Seems at some point our work here kinda broke down, and anything over
4+ digits in the input caused the first digits to gradually become
hidden.

Upping this scale factor to 10px/digit mostly fixes this issue, but now
we’re left with extra padding. Works for now :)

Suggest we eventually update this to some kind of an `em` or `rem`
inputWidth unit, so we can nail the width perfectly every time, and
allow it to scale with whatever font size is declared in CSS.