Reorder targets by using icon handle
#1014

Icon selector on the left is used to drag the entire target. The
targetName itself will be clickable, to navigate (and scroll) to the
expanded plan.

React-DnD has another source connector called `connectDragPreview`, to
tell the UI that it’s the whole target that drags, but only the icon as
the drag source.

Added some basic styling, not married to it yet…