Move shiftSummaries button to MainMenu item
#751

Was able to fit this into the MainMenu by slightly shrinking the height
of the menuItems (eventually we’ll want to refactor the styles here to
allow for more options, but this works for now)

ManagerLayer needed a little shim to ignore the module, which is just a
dialog (not a layer module).

Also updated Changelog.