Clicking an inactive target/section prompts SectionsView to expand it
#1014

Since sectionsView was put together so each group of inactive
sections/targets has its own state (for whether it’s open or not), it
made selectively expanding them from the top-level a bit of a headache.
It’d take a much broader refactor to pull up the state.

So, what happens:
1. Clicks inactive target in Reorder view
2. Regular plan view shows
3. Expand the deact/compl section group (if necessary)
4. Expand its internal deact/compl target group

This would all feel a bit more natural if it was combined into 1 single
view (expand / contract), would be a good >v2 upgrade!