Refactor scrolling to Async series, add highlighting styles
#1014

- Clicking a target will also make it the `selectedTargetId`, so
rightPane updates to show history. This just seems logical, considering
you’re clicking the target. Plus it provides a highlighted destination
for the scroll for free!

- Still playing with topOffset for scrolling to sections, added an
option to set the `additionalOffset`

- `.sectionName` briefly gets a `.highlight`class, which for now just
animates a dot on the left, and changes the color for a second. I’d
prefer to flash a background color, but it looks ugly in our current
layout.