Change "Reorder" to "Outline", move 1 button left
#1014

I experimented with the button ordering I had in mind during last
meeting - but it didn’t turn out looking as good as I’d hoped.

Turned out that just pushing it 1 button to the left looks way better.
When hasChanges, it’s separated from the “Save/Discard” buttons by the
disabled Print button. This also makes it a bit easier to find with the
eye after you’ve reordered something.

Used to not be clickable if hasChanges, but I’m going to add that
restriction directly into ReorderPlanView, so the user can still view
it - just not change.