Change short date format from 'MMM D' to 'D/M/YY' in Config
#966

DD/MM/YY was still too long (8 characters, versus “Mar 5” at 5 ch), so
I opted to go with D/M/YY, which comes in as 6-8.

I’m inclined to say that the full DD/MM/YY is slightly more readable,
but it really takes up a lot of horizontal space. Will keep playing
with the flex & sizings to make it work…