Ensure changing status triggers reason textarea focus()
Found this while testing #1039 . The target title (behind the dialog
layer) was focused instead, so I’d be modifying the name in the
background when I start typing. Delaying the focus ensures we override
it.

I think this has something to do with clicking the target area, and it
triggering the focus. But in planTarget.coffee line # 192, we see it
should be ignoring buttons.

Ah, now I see this is actually #1040 !