Allow prognote detail view item description to be expanded
Getting the actual rendered size of an element is tricky, so this
actually does two renders whenever the item description changes or the
visibility toggle button state changes. The first render allows the item
description height to be measured to determine if the description is
too long to fit in the box. The second render updates the button text
and behaviour appropriately.